+++ Sources/KeyboardKit/Gestures/DragGestureHandler.swift
@@ -1,5 +1,5 @@
 //
-//  KeyboardGesture.swift
+//  DragGestureHandler.swift
 //  KeyboardKit
 //
 //  Created by Daniel Saidi on 2021-04-01.