+++ Sources/KeyboardKit/Gestures/Gestures+Defaults.swift
@@ -9,10 +9,10 @@
 import Foundation
 
 public extension Gestures {
-    
+
     /// This type defines default gesture values.
     struct Defaults {
-        
+
         /// The max time between two taps for them to count as a double tap, by default `0.2`.
         public static var doubleTapTimeout = 0.2