+++ Sources/KeyboardKit/Previews/KeyboardPreviews+Actions.swift
@@ -21,12 +21,12 @@ public extension KeyboardPreviews {
         
         public init() {}
         
-        public func canHandle(_ gesture: Gesture, on action: KeyboardAction) -> Bool { false }
+        public func canHandle(_ gesture: Keyboard.Gesture, on action: KeyboardAction) -> Bool { false }
         public func handle(_ action: KeyboardAction) {}
         public func handle(_ suggestion: Autocomplete.Suggestion) {}
-        public func handle(_ gesture: Gesture, on action: KeyboardAction) {}
+        public func handle(_ gesture: Keyboard.Gesture, on action: KeyboardAction) {}
         public func handleDrag(on action: KeyboardAction, from startLocation: CGPoint, to currentLocation: CGPoint) {}
-        public func triggerFeedback(for gesture: Gesture, on action: KeyboardAction) {}
+        public func triggerFeedback(for gesture: Keyboard.Gesture, on action: KeyboardAction) {}
         public func triggerAudioFeedback(_ feedback: Feedback.Audio) {}
         public func triggerHapticFeedback(_ feedback: Feedback.Haptic) {}
     }