+++ Sources/KeyboardKit/_Deprecated/Gestures+Deprecated.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public extension Gestures {
+
+    @available(*, deprecated, renamed: "Keyboard.Gesture")
+    typealias KeyboardGesture = Keyboard.Gesture
+}